Create MaxProfitByBuyingInOneDayAndSellingInAnother.cpp
This question is about finding max profit of a merchant who will buy a stock in a day and sell it another day. Our purpose is to find the day in which day he/she should buy the stock to make maximum profit.

You are given prices for a particular stock for the next n days, say p 1 , p 2 , ....p n . You want to buy
one stock on one of the upcoming days and sell it on a later day. Both buying and selling should
be done within the next n days. You want to find out which buying and selling days will maximize
your profit. If multiple of these options are possible print with the earliest buying date.
Input:
The first line contains number of days n. 1 <= n <= 10 8
The second line contains prices separated by space
Output:
The first line prints maximum profit possible.
The second line prints the earliest buy date possible such that profit is maximised.
Sample:
Input:
5
100 10 1110 5000 4000
Output:
4990
2

In the example above, the way to make maximum profit is to buy stock on day 2 and sell it on day 4.